+++ CodeStarterCamp_Week5/AuditionManager.swift
@@ -31,10 +31,22 @@ struct AuditionManager {
         }
     }
     
+    func examinationOfSuccessfulApplicants() {
+        if passedApplicantsList.isEmpty == true {
+            print("불합격 입니다.")
+        }
+        
+        for passedPerson in passedApplicantsList {
+            print("\(passedPerson.name)님 합격을 축하드립니다.")
+        }
+    }
+    
     mutating func announcePassedApplicants() {
         print("--- 합격자 명단---")
-        cast()
+        examinationOfSuccessfulApplicants()
         print("---------------")
     }
+    
+   
     
 }
