+++ CodeStarterCamp_Week5/BadPersonality.swift
@@ -1,0 +1,12 @@
+//
+//  BadPersonality.swift
+//  CodeStarterCamp_Week5
+//
+//  Created by 1 on 2023/03/23.
+//
+
+import Foundation
+
+protocol BadPersonality {
+    var frequancyOfCursing: Level { get set }
+}
