+++ CodeStarterCamp_Week5/Talent.swift
@@ -1,0 +1,14 @@
+//
+//  Talent.swift
+//  CodeStarterCamp_Week5
+//
+//  Created by 1 on 2023/03/23.
+//
+
+import Foundation
+
+protocol Talent {
+    var singing: Level { get set }
+    var dancing: Level { get set }
+    var acting: Level { get set }
+}
